cleanup

+++ app/config/schema.sql
@@ -1,7 +1,7 @@
 PRAGMA foreign_keys = ON;
 
 CREATE TABLE IF NOT EXISTS users (
-  id BIGINT PRIMARY KEY NOT NULL,
+  id INTEGER PRIMARY KEY NOT NULL,
   username TEXT NOT NULL UNIQUE,
   email TEXT NOT NULL UNIQUE,
   profile_picture TEXT,
@@ -26,8 +26,8 @@ INSERT OR IGNORE INTO items (name) VALUES ("post");
 INSERT OR IGNORE INTO items (name) VALUES ("comment");
 
 CREATE TABLE IF NOT EXISTS posts (
-  id BIGINT PRIMARY KEY NOT NULL,          
-  user_id BIGINT NOT NULL,        
+  id INTEGER PRIMARY KEY NOT NULL,          
+  user_id INTEGER NOT NULL,        
   title TEXT NOT NULL,
   content TEXT NOT NULL,
   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
@@ -44,16 +44,16 @@ INSERT OR IGNORE INTO categories (name) VALUES ("technology");
 INSERT OR IGNORE INTO categories (name) VALUES ("science");
 
 CREATE TABLE IF NOT EXISTS post_categories (
-  post_id BIGINT NOT NULL,        -- References posts.id
+  post_id INTEGER NOT NULL,        -- References posts.id
   category_id INTEGER NOT NULL,
   FOREIGN KEY (post_id) REFERENCES posts(id) ON DELETE CASCADE,
   FOREIGN KEY (category_id) REFERENCES categories(id)
 );
 
 CREATE TABLE IF NOT EXISTS comments (
-  id BIGINT PRIMARY KEY NOT NULL, 
-  post_id BIGINT NOT NULL,        
-  user_id BIGINT NOT NULL,        
+  id INTEGER PRIMARY KEY NOT NULL, 
+  post_id INTEGER NOT NULL,        
+  user_id INTEGER NOT NULL,        
   content TEXT NOT NULL,
   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (post_id) REFERENCES posts(id),
@@ -69,8 +69,8 @@ INSERT OR IGNORE INTO reactions (name) VALUES ("dislike");
 
 CREATE TABLE IF NOT EXISTS item_reactions (
   item_type INTEGER NOT NULL,
-  item_id BIGINT NOT NULL, 
-  user_id BIGINT NOT NULL,        
+  item_id INTEGER NOT NULL, 
+  user_id INTEGER NOT NULL,        
   reaction_id INTEGER NOT NULL,   
   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (item_type) REFERENCES items(id),
